refactor: extract shell provider infrastructure into shell_infra.py
Move prompt detection, marker setup, KNOWN_SHELLS, PromptMatch, and
all shell mode helpers out of shell.py into a new shell_infra.py.
shell.py becomes a slim AgentProvider implementation that re-exports
names for backward compatibility. Adds chat_first_command_path capability
flag to ProviderCapabilities so handlers can gate shell-specific behaviour
without string-matching the provider name.
